requirements.txt
scapy==2.5.0
colorama==0.4.6
python-dateutil==2.8.2
scapy-http==1.8.2
matplotlib==3.7.1
networkx==3.1
dnspython==2.4.2
reportlab==4.0.4
jinja2==3.1.2
requests==2.31.0
python-whois==0.8.0
pyOpenSSL==23.2.0
flask
numpy==1.26.4
scikit-learn 